docs: move prose to layout
