+++ designsystem/field/field.mdx
@@ -7,7 +7,6 @@ import * as stories from './field.stories';
 # Field
 - Legg klassen `field` på typisk en `<div>` rundt ledetekst og skjemafelt
 - Dette kobler automatisk sammen `<label>` og `<input>`, samt eventuelle element med `<p>` og klassen `validation`
-- **Merk:** automatisk oppkobling krever at du har importert javascript pakken: `import "@mattilsynet/design"`
 <Canvas of={stories.Default} />
 
 ## Obligatorisk felt
