+++ designsystem/fieldset/fieldset.mdx
@@ -34,7 +34,6 @@ import * as stories from './fieldset.stories';
 
 ## Feilmelding
 - Legg klassen `validation` på et direkte barn for å knytte feilmelding til alle inputs
-- **Merk:** automatisk oppkobling krever at du har importert javascript pakken: `import "@mattilsynet/design"`
 <Canvas of={stories.WithValidation} />
 
 <CssVariables component="fieldset" />
