+++ designsystem/heading/heading.module.css
@@ -1,0 +1,12 @@
+@layer mt.design {
+	.heading {
+		composes: ds-heading from "@digdir/designsystemet-css";
+
+		scroll-margin-block: 4em; /* Ensure scrolling to headings still have some comfortable spacing above */
+		text-wrap: pretty;
+
+		&[data-justify="center"] {
+			text-align: center;
+		}
+	}
+}
