+++ designsystem/index.mdx
@@ -134,7 +134,6 @@ Designsystemet tilbyr først og fremst en rekke CSS klassenavn som hjelper deg b
     > Dette er super lett å overstyre ved behov: din CSS vil alltid ha første prioritert - takket være teknikken [CSS @layer](https://developer.mozilla.org/en-US/docs/Web/CSS/@layer) 💚
 4. **Ta i bruk stiler**
     - **Størrelser kan endres** med `data-size="sm | md | lg"` på *alle* HTML elementer - også HTML elementer uten designsystem-klasse
-    - **Nøst elementer** som f.eks.  `<dialog className={styles.dialog}><div className={styles.grid}>`, heller enn å kombinere flere designsystem-klasser på samme element
     - **Klassenavn** hentes fra  `styles` objektet:
 
     <u-tabs class={styles.tabs} data-size="sm">
