+++ designsystem/layout/layout.module.css
@@ -130,4 +130,61 @@
 	.grid[data-items="700"] { --mtds-grid-size: calc(1rem / 16 * 700) }
 	.grid[data-items="750"] { --mtds-grid-size: calc(1rem / 16 * 750) }
 	.grid[data-items="800"] { --mtds-grid-size: calc(1rem / 16 * 800) }
+
+
+	/**
+	 * Prose
+	 */
+	.prose > * {
+		margin-block: var(--ds-size-6);
+		max-width: 65ch;
+	}
+	.prose > :where(p, ul, ol, h1, h2, h3, h4, h5, h6) {
+		margin-block: 1em 0;
+	}
+	.prose > :where(ul, ol):not([class]) {
+		padding-left: .9em;
+
+		& li {
+			margin-top: var(--ds-size-2);
+		}
+		& > li:first-child {
+			margin-top: 0;
+		}
+	}
+
+	.prose > :where(h1, h2, h3, h4, h5, h6) {
+		&[data-size="lg"] {
+			margin-top: var(--ds-size-10);
+
+			& + p,
+			& + ul,
+			& + ol {
+				margin-top: var(--ds-size-6);
+			}
+		}
+		&[data-size="md"] {
+			margin-top: var(--ds-size-10);
+
+			& + p,
+			& + ul,
+			& + ol {
+				margin-top: var(--ds-size-2);
+			}
+		}
+		&[data-size="sm"] {
+			margin-top: var(--ds-size-6);
+
+			& + p {
+				margin-top: var(--ds-size-1);
+			}
+			& + ul,
+			& + ol {
+				margin-top: var(--ds-size-2);
+			}
+		}
+	}
+	.prose > :first-child {
+		margin-block: 0;
+	}
 }
