+++ designsystem/overview.mdx
@@ -14,6 +14,7 @@ import * as Divider from './divider/divider.stories.tsx';
 import * as Errorsummary from './errorsummary/errorsummary.stories.tsx';
 import * as Field from './field/field.stories.tsx';
 import * as Fieldset from './fieldset/fieldset.stories.tsx';
+import * as Heading from './heading/heading.stories.tsx';
 import * as Helptext from './helptext/helptext.stories.tsx';
 import * as Input from './input/input.stories.tsx';
 import * as Layout from './layout/layout.stories.tsx';
@@ -27,11 +28,10 @@ import * as Table from './table/table.stories.tsx';
 import * as Tabs from './tabs/tabs.stories.tsx';
 import * as Tag from './tag/tag.stories.tsx';
 import * as Tooltip from './tooltip/tooltip.stories.tsx';
-import * as Typography from './typography/typography.stories.tsx';
 import * as Validation from './validation/validation.stories.tsx';
 
 <Meta title="Designsystem/Komponenter" />
 
 # Komponenter
 
-<Overview items={[Alert, Avatar, Badge, Breadcrumbs, Button, Card, Chip, Details, Dialog, Divider, Errorsummary, Field, Fieldset, Helptext, Input, Layout, Link, Logo, Pagination, Popover, Skeleton, Spinner, Table, Tabs, Tag, Tooltip, Typography, Validation]} />
+<Overview items={[Alert, Avatar, Badge, Breadcrumbs, Button, Card, Chip, Details, Dialog, Divider, Errorsummary, Field, Fieldset, Heading, Helptext, Input, Layout, Link, Logo, Pagination, Popover, Skeleton, Spinner, Table, Tabs, Tag, Tooltip, Validation]} />
