+++ designsystem/popover/popover.mdx
@@ -8,7 +8,6 @@ import * as stories from './popover.stories';
 - Legg klassen `popover`, attributten `popover` og `id` på `<div>` eller `<menu>`
 - Legg `popovertarget="ID"` på typisk `<button>` som trigger element
 - Plassiering vil automatisk tilpasse seg scroll og tilgjengelig plass
-- **Merk:** automatisk posisjonering krever at du har importert javascript pakken: `import "@mattilsynet/design"`
 <Canvas of={stories.Default} />
 
 ## Nedtrekksmeny / Dropdown
