+++ designsystem/react.tsx
@@ -11,6 +11,7 @@ export * from "./divider/divider";
 export * from "./errorsummary/errorsummary";
 export * from "./field/field";
 export * from "./fieldset/fieldset";
+export * from "./heading/heading";
 export * from "./helptext/helptext";
 export * from "./input/input";
 export * from "./layout/layout";
@@ -23,5 +24,4 @@ export * from "./spinner/spinner";
 export * from "./table/table";
 export * from "./tabs/tabs";
 export * from "./tag/tag";
-export * from "./typography/typography";
 export * from "./validation/validation";
