--- designsystem/typography/typography.module.css
@@ -1,64 +1,0 @@
-@layer mt.design {
-	.heading {
-		composes: ds-heading from "@digdir/designsystemet-css";
-
-		scroll-margin-block: 4em; /* Ensure scrolling to headings still have some comfortable spacing above */
-		text-wrap: pretty;
-
-		&[data-justify="center"] {
-			text-align: center;
-		}
-	}
-	.prose > * {
-		margin-block: var(--ds-size-6);
-		max-width: 65ch;
-	}
-	.prose > :where(p, ul, ol, h1, h2, h3, h4, h5, h6) {
-		margin-block: 1em 0;
-	}
-	.prose > :where(ul, ol):not([class]) {
-		padding-left: .9em;
-
-		& li {
-			margin-top: var(--ds-size-2);
-		}
-		& > li:first-child {
-			margin-top: 0;
-		}
-	}
-
-	.prose > :where(h1, h2, h3, h4, h5, h6) {
-		&[data-size="lg"] {
-			margin-top: var(--ds-size-10);
-
-			& + p,
-			& + ul,
-			& + ol {
-				margin-top: var(--ds-size-6);
-			}
-		}
-		&[data-size="md"] {
-			margin-top: var(--ds-size-10);
-
-			& + p,
-			& + ul,
-			& + ol {
-				margin-top: var(--ds-size-2);
-			}
-		}
-		&[data-size="sm"] {
-			margin-top: var(--ds-size-6);
-
-			& + p {
-				margin-top: var(--ds-size-1);
-			}
-			& + ul,
-			& + ol {
-				margin-top: var(--ds-size-2);
-			}
-		}
-	}
-	.prose > :first-child {
-		margin-block: 0;
-	}
-}
